requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-jose[cryptography]
python-multipart
requests
